+++ lib/list_ext.dart
@@ -1,3 +1,4 @@
 library list_ext;
 
 export 'src/interable_extensions.dart';
+export 'src/list_extensions.dart';