+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Method `joinOf()` - get string value for each element and concatenates it with separator.
 * Method `chunks()` - splits into chunks (shortcut for quiver `partition()`).
 * Method `toMap()` - create `Map` from iterable (shortcut for `Map.fromIterable()`).
+* Extensions for `List` and method `copyWith()` - copy current list with adding element at the end of it.
 
 ## [0.1.1] - 2020-03-27